+++ backend/src/main/java/com/dubu/backend/member/api/PlaceController.java
@@ -2,7 +2,7 @@ package com.dubu.backend.member.api;
 
 import com.dubu.backend.global.domain.SuccessResponse;
 import com.dubu.backend.member.application.PlaceService;
-import com.dubu.backend.member.dto.PlaceSearchResponse;
+import com.dubu.backend.member.dto.response.PlaceSearchResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
