+++ backend/src/main/java/com/dubu/backend/member/application/MemberService.java
@@ -5,9 +5,9 @@ import com.dubu.backend.member.domain.Member;
 import com.dubu.backend.member.domain.MemberCategory;
 import com.dubu.backend.member.domain.enums.AddressType;
 import com.dubu.backend.member.domain.enums.Status;
-import com.dubu.backend.member.dto.MemberOnboardingRequest;
-import com.dubu.backend.member.dto.MemberSavedAddressResponse;
-import com.dubu.backend.member.dto.MemberStatusResponse;
+import com.dubu.backend.member.dto.request.MemberOnboardingRequest;
+import com.dubu.backend.member.dto.response.MemberSavedAddressResponse;
+import com.dubu.backend.member.dto.response.MemberStatusResponse;
 import com.dubu.backend.member.exception.MemberNotFoundException;
 import com.dubu.backend.member.exception.MemberSavedAddressNotFoundException;
 import com.dubu.backend.member.infra.repository.AddressRepository;
